Переработка логики запросов фильмов и обработки жанров подчиненных фильму
Сейчас для каждого фильма происходит отдельный запрос для жанров. Такая ситуация не совсем корректная, так как может сильно нагрузить приложение, эта называется проблема "N+1". Лучше оптимизировать запрос, чтобы он сразу искал для всех фильмов жанры. В данном классе несколько аналогичных недочетов

+++ src/main/java/ru/yandex/practicum/filmorete/service/ServiceFilm.java
@@ -78,18 +78,14 @@ public class ServiceFilm {
                     film.getName(), film.getDescription(),
                     film.getReleaseDate(), film.getDuration()
             );
-        } else {
-            throw new ExceptionNotFoundFilmStorage(VALID_ERROR_FILM_ID_NOT_IN_COLLECTIONS);
-        }
+        } else throw new ExceptionNotFoundFilmStorage(VALID_ERROR_FILM_ID_NOT_IN_COLLECTIONS);
         List<TotalGenreFilm> totalGenreFilms = totalGenreFilmDao.findAllTotalGenreFilm(film.getId());
-        if (!totalGenreFilms.isEmpty()) {
-            totalGenreFilmDao.deleteAllFilmId(film.getId());
-            if (film.getGenres() != null) {
-                for (Genre genreFilm : film.getGenres()) {
-                    Optional<TotalGenreFilm> totalGenreFilm = totalGenreFilmDao.findTotalGenreFilm(film.getId(), genreFilm.getId());
-                    if (totalGenreFilm.isEmpty()) {
-                        totalGenreFilmDao.insert(film.getId(), genreFilm.getId());
-                    }
+        if (!totalGenreFilms.isEmpty()) totalGenreFilmDao.deleteAllFilmId(film.getId());
+        if (film.getGenres() != null) {
+            for (Genre genreFilm : film.getGenres()) {
+                Optional<TotalGenreFilm> totalGenreFilm = totalGenreFilmDao.findTotalGenreFilm(film.getId(), genreFilm.getId());
+                if (totalGenreFilm.isEmpty()) {
+                    totalGenreFilmDao.insert(film.getId(), genreFilm.getId());
                 }
             }
         }
